factor out common pybullet code into a new base class (#718)
* factor out common pybullet code into a new base class

* revert

* yapf

* feedback

* yapf

* nit
