+++ src/envs/pybullet_blocks.py
@@ -1,7 +1,7 @@
 """A PyBullet version of Blocks."""
 
 import logging
-from typing import Callable, ClassVar, Dict, List, Optional, Sequence, Tuple
+from typing import Callable, ClassVar, Dict, List, Sequence, Tuple
 
 import numpy as np
 import pybullet as p
@@ -9,172 +9,113 @@ from gym.spaces import Box
 
 from predicators.src import utils
 from predicators.src.envs.blocks import BlocksEnv
-from predicators.src.envs.pybullet_robots import \
+from predicators.src.envs.pybullet_env import PyBulletEnv, _PyBulletState, \
+    create_pybullet_block
+from predicators.src.envs.pybullet_robots import _SingleArmPyBulletRobot, \
     create_single_arm_pybullet_robot
-from predicators.src.envs.pybullet_utils import _PyBulletState
 from predicators.src.settings import CFG
-from predicators.src.structs import Action, Array, Image, Object, \
-    ParameterizedOption, Pose3D, State, Task, Type
+from predicators.src.structs import Array, Object, ParameterizedOption, \
+    Pose3D, State
 
 
-class PyBulletBlocksEnv(BlocksEnv):
+class PyBulletBlocksEnv(PyBulletEnv, BlocksEnv):
     """PyBullet Blocks domain."""
     # Parameters that aren't important enough to need to clog up settings.py
 
-    # General robot parameters.
-    _move_gain: ClassVar[float] = 1.0
-    _max_vel_norm: ClassVar[float] = 0.05
-    _grasp_offset_z: ClassVar[float] = 0.01
-    _place_offset_z: ClassVar[float] = 0.01
-    _grasp_tol: ClassVar[float] = 0.05
-    _move_to_pose_tol: ClassVar[float] = 0.0001
-    _finger_action_tol: ClassVar[float] = 0.0001
+    # Option parameters.
+    _offset_z: ClassVar[float] = 0.01
 
     # Table parameters.
     _table_pose: ClassVar[Pose3D] = (1.35, 0.75, 0.0)
     _table_orientation: ClassVar[Sequence[float]] = [0., 0., 0., 1.]
 
-    # Block parameters.
-    _block_orientation: ClassVar[Sequence[float]] = [0., 0., 0., 1.]
-    _block_mass: ClassVar[float] = 0.5
-    _block_friction: ClassVar[float] = 1.2
-    _block_colors: ClassVar[Sequence[Tuple[float, float, float, float]]] = [
-        (0.95, 0.05, 0.1, 1.),
-        (0.05, 0.95, 0.1, 1.),
-        (0.1, 0.05, 0.95, 1.),
-        (0.4, 0.05, 0.6, 1.),
-        (0.6, 0.4, 0.05, 1.),
-        (0.05, 0.04, 0.6, 1.),
-        (0.95, 0.95, 0.1, 1.),
-        (0.95, 0.05, 0.95, 1.),
-        (0.05, 0.95, 0.95, 1.),
-    ]
-    _out_of_view_xy: ClassVar[Sequence[float]] = [10.0, 10.0]
-
-    # Camera parameters.
-    _camera_distance: ClassVar[float] = 0.8
-    _camera_yaw: ClassVar[float] = 90.0
-    _camera_pitch: ClassVar[float] = -24
-    _camera_target: ClassVar[Pose3D] = (1.65, 0.75, 0.42)
-    _debug_text_position: ClassVar[Pose3D] = (1.65, 0.25, 0.75)
-
     def __init__(self) -> None:
         super().__init__()
 
         # Override options, keeping the types and parameter spaces the same.
+
+        ## Pick option
         types = self._Pick.types
         params_space = self._Pick.params_space
         self._Pick: ParameterizedOption = utils.LinearChainParameterizedOption(
             "Pick",
             [
-                # Creates a ParameterizedOption for moving to the position that
-                # has z equal to self.pick_z, and x and y equal to that of the
-                # block object parameter. In other words, move the end effector
-                # to high above the block in preparation for picking.
-                self._create_move_end_effector_relative_to_block_option(
-                    "MoveEndEffectorToAboveBlock", (0., 0., self.pick_z),
-                    ("rel", "rel", "abs"), types, params_space),
+                # Move to far above the block which we will grasp.
+                self._create_move_to_above_block_option(
+                    name="MoveEndEffectorToPreGrasp",
+                    z_func=lambda _: self.pick_z),
                 # Open fingers.
                 self._create_change_fingers_option(
-                    "OpenFingers", self.open_fingers, types, params_space),
+                    "OpenFingers", self.open_fingers, types, params_space,
+                    self._robot),
                 # Move down to grasp.
-                self._create_move_end_effector_relative_to_block_option(
-                    "MoveEndEffectorToGrasp", (0., 0., self._grasp_offset_z),
-                    ("rel", "rel", "rel"), types, params_space),
+                self._create_move_to_above_block_option(
+                    name="MoveEndEffectorToGrasp",
+                    z_func=lambda block_z: (block_z + self._offset_z)),
                 # Grasp.
                 self._create_change_fingers_option(
-                    "Grasp", self.closed_fingers, types, params_space),
-                # Move up.
-                self._create_move_end_effector_relative_to_block_option(
-                    "MoveEndEffectorToAboveBlock", (0., 0., self.pick_z),
-                    ("rel", "rel", "abs"), types, params_space),
+                    "Grasp", self.closed_fingers, types, params_space,
+                    self._robot),
+                # Move back up.
+                self._create_move_to_above_block_option(
+                    name="MoveEndEffectorBackUp",
+                    z_func=lambda _: self.pick_z),
             ])
 
+        ## Stack option
         types = self._Stack.types
         params_space = self._Stack.params_space
         self._Stack: ParameterizedOption = \
             utils.LinearChainParameterizedOption("Stack",
             [
                 # Move to above the block on which we will stack.
-                self._create_move_end_effector_relative_to_block_option(
-                    "MoveEndEffectorToAboveBlock", (0., 0., self.pick_z),
-                    ("rel", "rel", "abs"), types, params_space),
+                self._create_move_to_above_block_option(
+                    name="MoveEndEffectorToPreStack",
+                    z_func=lambda _: self.pick_z),
                 # Move down to place.
-                self._create_move_end_effector_relative_to_block_option(
-                    "MoveEndEffectorToPlace",
-                    (0., 0., self.block_size + self._place_offset_z),
-                    ("rel", "rel", "rel"), types, params_space),
+                self._create_move_to_above_block_option(
+                    name="MoveEndEffectorToStack",
+                    z_func=lambda block_z: (
+                        block_z + self.block_size + self._offset_z)),
                 # Open fingers.
-                self._create_change_fingers_option("OpenFingers",
-                    self.open_fingers, types, params_space),
-                # Move up.
-                self._create_move_end_effector_relative_to_block_option(
-                    "MoveEndEffectorToAboveBlock", (0., 0., self.pick_z),
-                    ("rel", "rel", "abs"), types, params_space),
+                self._create_change_fingers_option(
+                    "OpenFingers", self.open_fingers, types, params_space,
+                    self._robot),
+                # Move back up.
+                self._create_move_to_above_block_option(
+                    name="MoveEndEffectorBackUp",
+                    z_func=lambda _: self.pick_z),
             ])
 
+        ## PutOnTable option
         types = self._PutOnTable.types
         params_space = self._PutOnTable.params_space
+        place_z = self.table_height + self.block_size + self._offset_z
         self._PutOnTable: ParameterizedOption = \
             utils.LinearChainParameterizedOption("PutOnTable",
             [
                 # Move to above the table at the (x, y) where we will place.
-                self._create_move_end_effector_relative_to_table_option(
-                    "MoveEndEffectorToAboveTable", (0., 0., self.pick_z),
-                    ("rel", "rel", "abs"), types, params_space),
+                self._create_move_to_above_table_option(
+                    name="MoveEndEffectorToPrePutOnTable", z=self.pick_z),
                 # Move down to place.
-                self._create_move_end_effector_relative_to_table_option(
-                    "MoveEndEffectorToPlaceOnTable",
-                    (0., 0., self.block_size + self._place_offset_z),
-                    ("rel", "rel", "rel"), types, params_space),
+                self._create_move_to_above_table_option(
+                    name="MoveEndEffectorToPutOnTable", z=place_z),
                 # Open fingers.
-                self._create_change_fingers_option("OpenFingers",
-                    self.open_fingers, types, params_space),
-                # Move up.
-                self._create_move_end_effector_relative_to_table_option(
-                    "MoveEndEffectorToAboveTable", (0., 0., self.pick_z),
-                    ("rel", "rel", "abs"), types, params_space),
+                self._create_change_fingers_option(
+                    "OpenFingers", self.open_fingers, types, params_space,
+                    self._robot),
+                # Move back up.
+                self._create_move_to_above_table_option(
+                    name="MoveEndEffectorBackUp", z=self.pick_z),
             ])
 
         # We track the correspondence between PyBullet object IDs and Object
         # instances for blocks. This correspondence changes with the task.
         self._block_id_to_block: Dict[int, Object] = {}
 
-        # When a block is held, a constraint is created to prevent slippage.
-        self._held_constraint_id: Optional[int] = None
-        self._held_block_id: Optional[int] = None
-
-        # Set up all the static PyBullet content.
-        self._initialize_pybullet()
-
     def _initialize_pybullet(self) -> None:
-        """One-time initialization of pybullet assets."""
-        # Skip test coverage because GUI is too expensive to use in unit tests
-        # and cannot be used in headless mode.
-        if CFG.pybullet_use_gui:  # pragma: no cover
-            self._physics_client_id = p.connect(p.GUI)
-            p.resetDebugVisualizerCamera(
-                self._camera_distance,
-                self._camera_yaw,
-                self._camera_pitch,
-                self._camera_target,
-                physicsClientId=self._physics_client_id)
-        else:
-            self._physics_client_id = p.connect(p.DIRECT)
-
-        p.resetSimulation(physicsClientId=self._physics_client_id)
-
-        # Load plane.
-        p.loadURDF(utils.get_env_asset_path("urdf/plane.urdf"), [0, 0, -1],
-                   useFixedBase=True,
-                   physicsClientId=self._physics_client_id)
-
-        # Load and reset robot.
-        ee_home = (self.robot_init_x, self.robot_init_y, self.robot_init_z)
-        self._pybullet_robot = create_single_arm_pybullet_robot(
-            CFG.pybullet_robot, ee_home, self.open_fingers,
-            self.closed_fingers, self._finger_action_tol, self._max_vel_norm,
-            self._physics_client_id)
+        """Run super(), then handle blocks-specific initialization."""
+        super()._initialize_pybullet()
 
         # Load table.
         self._table_id = p.loadURDF(
@@ -219,56 +160,45 @@ class PyBulletBlocksEnv(BlocksEnv):
             p.addUserDebugText("*", [mid_x, mid_y, self.pick_z],
                                [1.0, 0.0, 0.0])
 
-        # Set gravity.
-        p.setGravity(0., 0., -10., physicsClientId=self._physics_client_id)
-
         # Create blocks. Note that we create the maximum number once, and then
-        # remove blocks from the workspace (teleporting them far away) based on
-        # the number involved in the state.
+        # later on, in reset_state(), we will remove blocks from the workspace
+        # (teleporting them far away) based on which ones are in the state.
         num_blocks = max(max(CFG.blocks_num_blocks_train),
                          max(CFG.blocks_num_blocks_test))
-        self._block_ids = [self._create_block(i) for i in range(num_blocks)]
+        self._block_ids = []
+        for i in range(num_blocks):
+            color = self._obj_colors[i % len(self._obj_colors)]
+            orientation = [0.0, 0.0, 0.0, 1.0]  # default
+            self._block_ids.append(
+                create_pybullet_block(color, self.block_size, self._obj_mass,
+                                      self._obj_friction, orientation,
+                                      self._physics_client_id))
 
-    @property
-    def action_space(self) -> Box:
-        return self._pybullet_robot.action_space
+    def _create_pybullet_robot(self) -> _SingleArmPyBulletRobot:
+        ee_home = (self.robot_init_x, self.robot_init_y, self.robot_init_z)
+        return create_single_arm_pybullet_robot(CFG.pybullet_robot, ee_home,
+                                                self.open_fingers,
+                                                self.closed_fingers,
+                                                self._finger_action_tol,
+                                                self._max_vel_norm,
+                                                self._physics_client_id)
+
+    def _extract_robot_state(self, state: State) -> Array:
+        return np.array([
+            state.get(self._robot, "pose_x"),
+            state.get(self._robot, "pose_y"),
+            state.get(self._robot, "pose_z"),
+            state.get(self._robot, "fingers")
+        ],
+                        dtype=np.float32)
 
     @classmethod
     def get_name(cls) -> str:
         return "pybullet_blocks"
 
-    def simulate(self, state: State, action: Action) -> State:
-        raise NotImplementedError("PyBulletBlocksEnv cannot simulate.")
-
-    def render_state(self,
-                     state: State,
-                     task: Task,
-                     action: Optional[Action] = None,
-                     caption: Optional[str] = None) -> List[Image]:
-        raise NotImplementedError("PyBulletBlocksEnv cannot render arbitrary "
-                                  "states.")
-
-    def reset(self, train_or_test: str, task_idx: int) -> State:
-        state = super().reset(train_or_test, task_idx)
-        self._reset_state(state)
-        # We could call self._get_state() here, but there could be small
-        # inconsistencies between that and the state expected as the initial
-        # train task state. Giving the expected initial state in this way
-        # leads to a tiny improvement in performance.
-        joint_state = list(self._pybullet_robot.initial_joint_values)
-        return _PyBulletState(state.data, simulator_state=joint_state)
-
     def _reset_state(self, state: State) -> None:
-        """Helper for reset and testing."""
-        # Tear down the old PyBullet scene.
-        if self._held_constraint_id is not None:
-            p.removeConstraint(self._held_constraint_id,
-                               physicsClientId=self._physics_client_id)
-            self._held_constraint_id = None
-        self._held_block_id = None
-
-        # Reset robot.
-        self._pybullet_robot.reset_state(state[self._robot])
+        """Run super(), then handle blocks-specific resetting."""
+        super()._reset_state(state)
 
         # Reset blocks based on the state.
         block_objs = state.get_objects(self._block_type)
@@ -276,12 +206,13 @@ class PyBulletBlocksEnv(BlocksEnv):
         for i, block_obj in enumerate(block_objs):
             block_id = self._block_ids[i]
             self._block_id_to_block[block_id] = block_obj
-            bx, by, bz, _ = state[block_obj]
+            bx = state.get(block_obj, "pose_x")
+            by = state.get(block_obj, "pose_y")
+            bz = state.get(block_obj, "pose_z")
             # Assume not holding in the initial state
             assert self._get_held_block(state) is None
             p.resetBasePositionAndOrientation(
-                block_id, [bx, by, bz],
-                self._block_orientation,
+                block_id, [bx, by, bz], [0.0, 0.0, 0.0, 1.0],
                 physicsClientId=self._physics_client_id)
 
         # For any blocks not involved, put them out of view.
@@ -291,8 +222,7 @@ class PyBulletBlocksEnv(BlocksEnv):
             block_id = self._block_ids[i]
             assert block_id not in self._block_id_to_block
             p.resetBasePositionAndOrientation(
-                block_id, [oov_x, oov_y, i * h],
-                self._block_orientation,
+                block_id, [oov_x, oov_y, i * h], [0.0, 0.0, 0.0, 1.0],
                 physicsClientId=self._physics_client_id)
 
         # Assert that the state was properly reconstructed.
@@ -304,150 +234,11 @@ class PyBulletBlocksEnv(BlocksEnv):
             logging.debug(reconstructed_state.pretty_str())
             raise ValueError("Could not reconstruct state.")
 
-    def _create_block(self, block_num: int) -> int:
-        """Returns the body ID."""
-        color = self._block_colors[block_num % len(self._block_colors)]
-
-        # The poses here are not important because they are overwritten by
-        # the state values when a task is reset. By default, we just stack all
-        # the blocks into one pile at the center of the table.
-        h = self.block_size
-        x = (self.x_lb + self.x_ub) / 2
-        y = (self.y_lb + self.y_ub) / 2
-        z = self.table_height + (0.5 * h) + (h * block_num)
-        pose = (x, y, z)
-        half_extents = [self.block_size / 2.] * 3
-
-        # Create the collision shape.
-        collision_id = p.createCollisionShape(
-            p.GEOM_BOX,
-            halfExtents=half_extents,
-            physicsClientId=self._physics_client_id)
-
-        # Create the visual_shape.
-        visual_id = p.createVisualShape(
-            p.GEOM_BOX,
-            halfExtents=half_extents,
-            rgbaColor=color,
-            physicsClientId=self._physics_client_id)
-
-        # Create the body.
-        block_id = p.createMultiBody(baseMass=self._block_mass,
-                                     baseCollisionShapeIndex=collision_id,
-                                     baseVisualShapeIndex=visual_id,
-                                     basePosition=pose,
-                                     baseOrientation=self._block_orientation,
-                                     physicsClientId=self._physics_client_id)
-        p.changeDynamics(
-            block_id,
-            linkIndex=-1,  # -1 for the base
-            lateralFriction=self._block_friction,
-            physicsClientId=self._physics_client_id)
-
-        return block_id
-
-    def render(
-            self,
-            action: Optional[Action] = None,
-            caption: Optional[str] = None) -> List[Image]:  # pragma: no cover
-        # Skip test coverage because GUI is too expensive to use in unit tests
-        # and cannot be used in headless mode.
-        del caption  # unused
-
-        if not CFG.pybullet_use_gui:
-            raise Exception(
-                "Rendering only works with GUI on. See "
-                "https://github.com/bulletphysics/bullet3/issues/1157")
-
-        view_matrix = p.computeViewMatrixFromYawPitchRoll(
-            cameraTargetPosition=self._camera_target,
-            distance=self._camera_distance,
-            yaw=self._camera_yaw,
-            pitch=self._camera_pitch,
-            roll=0,
-            upAxisIndex=2,
-            physicsClientId=self._physics_client_id)
-
-        width = CFG.pybullet_camera_width
-        height = CFG.pybullet_camera_height
-
-        proj_matrix = p.computeProjectionMatrixFOV(
-            fov=60,
-            aspect=float(width / height),
-            nearVal=0.1,
-            farVal=100.0,
-            physicsClientId=self._physics_client_id)
-
-        (_, _, px, _,
-         _) = p.getCameraImage(width=width,
-                               height=height,
-                               viewMatrix=view_matrix,
-                               projectionMatrix=proj_matrix,
-                               renderer=p.ER_BULLET_HARDWARE_OPENGL,
-                               physicsClientId=self._physics_client_id)
-
-        rgb_array = np.array(px)
-        rgb_array = rgb_array[:, :, :3]
-        return [rgb_array]
-
-    def step(self, action: Action) -> State:
-        # Send the action to the robot.
-        ee_delta = (action.arr[0], action.arr[1], action.arr[2])
-        f_delta = action.arr[3]
-        self._pybullet_robot.set_motors(ee_delta, f_delta)
-
-        # Step the simulation here before adding or removing constraints
-        # because detect_held_block() should use the updated state.
-        for _ in range(CFG.pybullet_sim_steps_per_action):
-            p.stepSimulation(physicsClientId=self._physics_client_id)
-
-        # If not currently holding something, and fingers are closing, check
-        # for a new grasp.
-        if self._held_constraint_id is None and \
-            f_delta < -self._finger_action_tol:
-            # Detect whether an object is held.
-            self._held_block_id = self._detect_held_block()
-            if self._held_block_id is not None:
-                # Create a grasp constraint.
-                base_link_to_world = np.r_[p.invertTransform(*p.getLinkState(
-                    self._pybullet_robot.robot_id,
-                    self._pybullet_robot.end_effector_id,
-                    physicsClientId=self._physics_client_id)[:2])]
-                world_to_obj = np.r_[p.getBasePositionAndOrientation(
-                    self._held_block_id,
-                    physicsClientId=self._physics_client_id)]
-                base_link_to_obj = p.invertTransform(*p.multiplyTransforms(
-                    base_link_to_world[:3], base_link_to_world[3:],
-                    world_to_obj[:3], world_to_obj[3:]))
-                self._held_constraint_id = p.createConstraint(
-                    parentBodyUniqueId=self._pybullet_robot.robot_id,
-                    parentLinkIndex=self._pybullet_robot.end_effector_id,
-                    childBodyUniqueId=self._held_block_id,
-                    childLinkIndex=-1,  # -1 for the base
-                    jointType=p.JOINT_FIXED,
-                    jointAxis=[0, 0, 0],
-                    parentFramePosition=[0, 0, 0],
-                    childFramePosition=base_link_to_obj[0],
-                    parentFrameOrientation=[0, 0, 0, 1],
-                    childFrameOrientation=base_link_to_obj[1],
-                    physicsClientId=self._physics_client_id)
-
-        # If placing, remove the grasp constraint.
-        if self._held_constraint_id is not None and \
-            f_delta > self._finger_action_tol:
-            p.removeConstraint(self._held_constraint_id,
-                               physicsClientId=self._physics_client_id)
-            self._held_constraint_id = None
-            self._held_block_id = None
-
-        self._current_state = self._get_state()
-        return self._current_state.copy()
-
     def _get_state(self) -> State:
         """Create a State based on the current PyBullet state.
 
         Note that in addition to the state inside PyBullet itself, this
-        uses self._block_id_to_block and self._held_block_id. As long as
+        uses self._block_id_to_block and self._held_obj_id. As long as
         the PyBullet internal state is only modified through reset() and
         step(), these all should remain in sync.
         """
@@ -461,7 +252,7 @@ class PyBulletBlocksEnv(BlocksEnv):
         for block_id, block in self._block_id_to_block.items():
             (bx, by, bz), _ = p.getBasePositionAndOrientation(
                 block_id, physicsClientId=self._physics_client_id)
-            held = (block_id == self._held_block_id)
+            held = (block_id == self._held_obj_id)
             # pose_x, pose_y, pose_z, held
             state_dict[block] = np.array([bx, by, bz, held], dtype=np.float32)
 
@@ -472,191 +263,59 @@ class PyBulletBlocksEnv(BlocksEnv):
 
         return state
 
-    def _detect_held_block(self) -> Optional[int]:
-        """Return the PyBullet object ID of the held object if one exists.
+    def _get_object_ids_for_held_check(self) -> List[int]:
+        return sorted(self._block_id_to_block)
 
-        If multiple blocks are within the grasp tolerance, return the
-        one that is closest.
+    def _create_move_to_above_block_option(
+            self, name: str, z_func: Callable[[float],
+                                              float]) -> ParameterizedOption:
+        """Creates a ParameterizedOption for moving to a pose above that of the
+        block argument.
+
+        The parameter z_func maps the block's z position to the target z
+        position.
         """
-        expected_finger_normals = {
-            self._pybullet_robot.left_finger_id: np.array([0., 1., 0.]),
-            self._pybullet_robot.right_finger_id: np.array([0., -1., 0.]),
-        }
-        closest_held_block = None
-        closest_held_block_dist = float("inf")
-        for block_id in self._block_id_to_block:
-            for finger_id, expected_normal in expected_finger_normals.items():
-                # Find points on the block that are within grasp_tol distance
-                # of the finger. Note that we use getClosestPoints instead of
-                # getContactPoints because we still want to consider the block
-                # held even if there is a tiny distance between the fingers and
-                # the block.
-                closest_points = p.getClosestPoints(
-                    bodyA=self._pybullet_robot.robot_id,
-                    bodyB=block_id,
-                    distance=self._grasp_tol,
-                    linkIndexA=finger_id,
-                    physicsClientId=self._physics_client_id)
-                for point in closest_points:
-                    # If the contact normal is substantially different from
-                    # the expected contact normal, this is probably a block
-                    # on the outside of the fingers, rather than the inside.
-                    # A perfect score here is 1.0 (normals are unit vectors).
-                    contact_normal = point[7]
-                    if expected_normal.dot(contact_normal) < 0.5:
-                        continue
-                    # Handle the case where multiple blocks pass this check
-                    # by taking the closest one. This should be rare, but it
-                    # can happen when two blocks are stacked and the robot is
-                    # unstacking the top one.
-                    contact_distance = point[8]
-                    if contact_distance < closest_held_block_dist:
-                        closest_held_block = block_id
-                        closest_held_block_dist = contact_distance
-        return closest_held_block
-
-    ########################## Parameterized Options ##########################
-
-    def _create_move_end_effector_to_pose_option(
-        self, name: str, types: Sequence[Type], params_space: Box,
-        get_current_and_target_pose: Callable[[State, Sequence[Object], Array],
-                                              Tuple[Pose3D, Pose3D]]
-    ) -> ParameterizedOption:
-        """Creates a ParameterizedOption for moving the end effector to a
-        target pose, given a function that takes in the current state and
-        objects and returns the current pose and target pose of the robot."""
-
-        def _policy(state: State, memory: Dict, objects: Sequence[Object],
-                    params: Array) -> Action:
-            del memory  # unused
-            current, target = get_current_and_target_pose(
-                state, objects, params)
-            return self._get_end_effector_move_toward_waypoint_action(
-                current, target)
-
-        def _terminal(state: State, memory: Dict, objects: Sequence[Object],
-                      params: Array) -> bool:
-            del memory  # unused
-            current, target = get_current_and_target_pose(
-                state, objects, params)
-            squared_dist = np.sum(np.square(np.subtract(current, target)))
-            return squared_dist < self._move_to_pose_tol
-
-        return ParameterizedOption(name,
-                                   types=types,
-                                   params_space=params_space,
-                                   policy=_policy,
-                                   initiable=lambda _1, _2, _3, _4: True,
-                                   terminal=_terminal)
-
-    def _create_move_end_effector_relative_to_block_option(
-            self, name: str, rel_or_abs_target_pose: Pose3D,
-            rel_or_abs: Tuple[str, str, str], types: Sequence[Type],
-            params_space: Box) -> ParameterizedOption:
-        """Creates a ParameterizedOption for moving the end effector to a
-        target pose relative to a block.
-
-        Each of the three dimensions of the target pose can be specified
-        relatively or absolutely. The rel_or_abs argument indicates
-        whether each dimension is relative ("rel") or absolute ("abs").
-        """
-        assert types == [self._robot_type, self._block_type]
+        types = [self._robot_type, self._block_type]
+        params_space = Box(0, 1, (0, ))
 
         def _get_current_and_target_pose(
                 state: State, objects: Sequence[Object],
                 params: Array) -> Tuple[Pose3D, Pose3D]:
-            del params  # not used
+            assert not params
             robot, block = objects
-            current_pose = (state[robot][0], state[robot][1], state[robot][2])
-            block_pose = (state[block][0], state[block][1], state[block][2])
-            target_pose = self._convert_rel_abs_to_abs(block_pose,
-                                                       rel_or_abs_target_pose,
-                                                       rel_or_abs)
+            current_pose = (state.get(robot,
+                                      "pose_x"), state.get(robot, "pose_y"),
+                            state.get(robot, "pose_z"))
+            target_pose = (state.get(block,
+                                     "pose_x"), state.get(block, "pose_y"),
+                           z_func(state.get(block, "pose_z")))
             return current_pose, target_pose
 
         return self._create_move_end_effector_to_pose_option(
             name, types, params_space, _get_current_and_target_pose)
 
-    def _create_move_end_effector_relative_to_table_option(
-            self, name: str, rel_or_abs_target_pose: Pose3D,
-            rel_or_abs: Tuple[str, str, str], types: Sequence[Type],
-            params_space: Box) -> ParameterizedOption:
-        """Creates a ParameterizedOption for moving the end effector to a
-        target pose relative to the table.
+    def _create_move_to_above_table_option(self, name: str,
+                                           z: float) -> ParameterizedOption:
+        """Creates a ParameterizedOption for moving to a pose above that of the
+        table.
 
-        Each of the three dimensions of the target pose can be specified
-        relatively or absolutely. The rel_or_abs argument indicates
-        whether each dimension is relative ("rel") or absolute ("abs").
+        The z position of the target pose must be provided.
         """
+        types = [self._robot_type]
+        params_space = Box(0, 1, (2, ))
 
         def _get_current_and_target_pose(
                 state: State, objects: Sequence[Object],
                 params: Array) -> Tuple[Pose3D, Pose3D]:
             robot, = objects
+            current_pose = (state.get(robot,
+                                      "pose_x"), state.get(robot, "pose_y"),
+                            state.get(robot, "pose_z"))
             # De-normalize parameters to actual table coordinates.
             x_norm, y_norm = params
-            x = self.x_lb + (self.x_ub - self.x_lb) * x_norm
-            y = self.y_lb + (self.y_ub - self.y_lb) * y_norm
-            current_pose = (state[robot][0], state[robot][1], state[robot][2])
-            table_pose = (x, y, self.table_height)
-            target_pose = self._convert_rel_abs_to_abs(table_pose,
-                                                       rel_or_abs_target_pose,
-                                                       rel_or_abs)
+            target_pose = (self.x_lb + (self.x_ub - self.x_lb) * x_norm,
+                           self.y_lb + (self.y_ub - self.y_lb) * y_norm, z)
             return current_pose, target_pose
 
         return self._create_move_end_effector_to_pose_option(
             name, types, params_space, _get_current_and_target_pose)
-
-    def _create_change_fingers_option(
-            self, name: str, target_val: float, types: Sequence[Type],
-            params_space: Box) -> ParameterizedOption:
-        """Creates a ParameterizedOption for changing the robot fingers."""
-
-        assert types[0] == self._robot_type
-
-        def _policy(state: State, memory: Dict, objects: Sequence[Object],
-                    params: Array) -> Action:
-            del memory, objects, params  # unused
-            current_val = state.get(self._robot, "fingers")
-            f_delta = target_val - current_val
-            f_delta = np.clip(f_delta, self.action_space.low[3],
-                              self.action_space.high[3])
-            return Action(np.array([0., 0., 0., f_delta], dtype=np.float32))
-
-        def _terminal(state: State, memory: Dict, objects: Sequence[Object],
-                      params: Array) -> bool:
-            del memory, objects, params  # unused
-            current_val = state.get(self._robot, "fingers")
-            squared_dist = (target_val - current_val)**2
-            return squared_dist < self._grasp_tol
-
-        return ParameterizedOption(name,
-                                   types=types,
-                                   params_space=params_space,
-                                   policy=_policy,
-                                   initiable=lambda _1, _2, _3, _4: True,
-                                   terminal=_terminal)
-
-    def _get_end_effector_move_toward_waypoint_action(
-            self, ee_pose: Pose3D, target_pose: Pose3D) -> Action:
-        action = self._move_gain * np.subtract(target_pose, ee_pose)
-        action_norm = np.linalg.norm(action)  # type: ignore
-        if action_norm > self._max_vel_norm:
-            action = action * self._max_vel_norm / action_norm
-        action = np.r_[action, 0.0].astype(np.float32)
-        assert self.action_space.contains(action)
-        return Action(action)
-
-    @staticmethod
-    def _convert_rel_abs_to_abs(current_pose: Pose3D, rel_or_abs_pose: Pose3D,
-                                rel_or_abs: Tuple[str, str, str]) -> Pose3D:
-        abs_pos = []
-        assert len(current_pose) == len(rel_or_abs_pose) == len(rel_or_abs)
-        for i in range(len(current_pose)):
-            if rel_or_abs[i] == "rel":
-                pose_i = current_pose[i] + rel_or_abs_pose[i]
-            else:
-                assert rel_or_abs[i] == "abs"
-                pose_i = rel_or_abs_pose[i]
-            abs_pos.append(pose_i)
-        return (abs_pos[0], abs_pos[1], abs_pos[2])
