+++ src/envs/pybullet_robots.py
@@ -8,8 +8,7 @@ import pybullet as p
 from gym.spaces import Box
 
 from predicators.src import utils
-from predicators.src.envs.pybullet_utils import get_kinematic_chain, \
-    inverse_kinematics
+from predicators.src.settings import CFG
 from predicators.src.structs import Array, Pose3D
 
 
@@ -307,3 +306,105 @@ def create_single_arm_pybullet_robot(
                                   finger_action_tol, max_vel_norm,
                                   physics_client_id)
     raise NotImplementedError(f"Unrecognized robot name: {robot_name}.")
+
+
+def get_kinematic_chain(robot: int, end_effector: int,
+                        physics_client_id: int) -> List[int]:
+    """Get all of the free joints from robot base to end effector.
+
+    Includes the end effector.
+    """
+    kinematic_chain = []
+    while end_effector > -1:
+        joint_info = p.getJointInfo(robot,
+                                    end_effector,
+                                    physicsClientId=physics_client_id)
+        if joint_info[3] > -1:
+            kinematic_chain.append(end_effector)
+        end_effector = joint_info[-1]
+    return kinematic_chain
+
+
+def inverse_kinematics(
+    robot: int,
+    end_effector: int,
+    target_position: Sequence[float],
+    target_orientation: Sequence[float],
+    joints: Sequence[int],
+    physics_client_id: int,
+    validate: bool = True,
+) -> List[float]:
+    """Runs IK and returns joint values for the given (free) joints.
+
+    If validate is True, the PyBullet IK solver is called multiple
+    times, resetting the robot state each time, until the target
+    position is reached. If the target position is not reached after a
+    maximum number of iters, an exception is raised.
+    """
+    # Figure out which joint each dimension of the return of IK corresponds to.
+    free_joints = []
+    num_joints = p.getNumJoints(robot, physicsClientId=physics_client_id)
+    for idx in range(num_joints):
+        joint_info = p.getJointInfo(robot,
+                                    idx,
+                                    physicsClientId=physics_client_id)
+        if joint_info[3] > -1:
+            free_joints.append(idx)
+    assert set(joints).issubset(set(free_joints))
+
+    # Record the initial state of the joints so that we can reset them after.
+    if validate:
+        initial_joint_states = p.getJointStates(
+            robot, free_joints, physicsClientId=physics_client_id)
+        assert len(initial_joint_states) == len(free_joints)
+
+    # Running IK once is often insufficient, so we run it multiple times until
+    # convergence. If it does not converge, an error is raised.
+    convergence_tol = CFG.pybullet_ik_tol
+    for _ in range(CFG.pybullet_max_ik_iters):
+        free_joint_vals = p.calculateInverseKinematics(
+            robot,
+            end_effector,
+            target_position,
+            targetOrientation=target_orientation,
+            physicsClientId=physics_client_id)
+        assert len(free_joints) == len(free_joint_vals)
+        if not validate:
+            break
+        # Update the robot state and check if the desired position and
+        # orientation are reached.
+        for joint, joint_val in zip(free_joints, free_joint_vals):
+            p.resetJointState(robot,
+                              joint,
+                              targetValue=joint_val,
+                              physicsClientId=physics_client_id)
+        ee_link_state = p.getLinkState(robot,
+                                       end_effector,
+                                       computeForwardKinematics=True,
+                                       physicsClientId=physics_client_id)
+        position = ee_link_state[4]
+        # Note: we are checking positions only for convergence.
+        if np.allclose(position, target_position, atol=convergence_tol):
+            break
+    else:
+        raise Exception("Inverse kinematics failed to converge.")
+
+    # Reset the joint states to their initial values to avoid modifying the
+    # PyBullet internal state.
+    if validate:
+        for joint, joint_state in zip(free_joints, initial_joint_states):
+            position, velocity, _, _ = joint_state
+            p.resetJointState(robot,
+                              joint,
+                              targetValue=position,
+                              targetVelocity=velocity,
+                              physicsClientId=physics_client_id)
+
+    # Order the found free_joint_vals based on the requested joints.
+    joint_vals = []
+    for joint in joints:
+        free_joint_idx = free_joints.index(joint)
+        joint_val = free_joint_vals[free_joint_idx]
+        joint_vals.append(joint_val)
+
+    return joint_vals
